(fix): keep public double type and finish query-lib hook wiring
Inline attribute creation advertised `double`; switching the whitelist
to `float` broke the documented API. Shared-table CLI databases also
skipped Tenancy, organization project create wrapped typed schema as
Document, and document-count metrics never reached getDatabasesDB.

+++ src/Appwrite/Migration/Version/V18.php
@@ -57,7 +57,7 @@ class V18 extends Migration
                 $collectionTable = "{$databaseTable}_collection_{$collection->getSequence()}";
 
                 foreach ($collection['attributes'] ?? [] as $attribute) {
-                    if ($attribute['type'] !== ColumnType::Float->value) {
+                    if ($attribute['type'] !== ColumnType::Double->value) {
                         continue;
                     }
                     $this->changeAttributeInternalType($collectionTable, $attribute['key'], 'DOUBLE');
@@ -88,7 +88,7 @@ class V18 extends Migration
             Console::log("Migrating Collection \"{$id}\"");
 
             foreach ($collection['attributes'] ?? [] as $attribute) {
-                if ($attribute['type'] !== ColumnType::Float->value) {
+                if ($attribute['type'] !== ColumnType::Double->value) {
                     continue;
                 }
                 $this->changeAttributeInternalType($id, $attribute['$id'], 'DOUBLE');
